feat(makeall.sh): store generated models in the ./output directory
Modified the `./makeall.sh` script to store all generated models
within the `./output` directory of the project. As part of this
feature, the `--output-dir` option was implemented in the python
commands called by `makeall.sh`. This new option allows users to
specify a custom output directory for the generated models.