feat(download_taesd_models.sh): --wget option and improved download sequence
Enhanced the sequence of downloads and added improved user prompts
in the `./download_taesd_model.sh` script.

As a significant improvement, the script now includes an option to
change the command used for downloading files. By default, the script
uses curl, but if the `--wget` option is utilized, the script will
switch to using wget.